Added login flow
**1) What is being done?**

To add an app to one or more sub-accounts we need to be able to login and get the API Key for the agency user. The locations api requires the use of an API Key that is returned when the user logs into the GHL [app](https://app.gohighlevel.com/) & not the [Marketplace](https://marketplace.gohighlevel.com) app.

The complete login involves:
1. Verifying the email+password
2. Getting an OTP
3. Validating the OTP
to get the API Key.

**2) What has changed?**

* Updated somme angular dependencies
* Added the code for handling the complete GHL Web app login flow

**3) Any concern?**

None

**4) Links to design/ other documentation**

N/A

**5) Testing**

Manual testing of the UI