Landing page for showing all GHL Marketplace Apps
**1) What is being done?**

When the user first navigates to the app, they should be shown a list of all apps fromm the GHL Marketplace. They should also be able to search for an app and click on it to go to the detail page, from where they can add it to 1 or more sub-accounts.

The search will work on these properties of the app:
1. companyName
2. name
3. description

**2) What has changed?**

* Updated routing to make the marketplace page the default landing page
* Added code to show a list of apps & the ability to search for one from the list

**3) Any concern?**

None

**4) Links to design/ other documentation**

N/A

**5) Testing**

Manual testing of the UI