Updated GHL App model
**1) What is being done?**

The GHL API for getting app details has 2 variations:
1. Public marketplace apps
2. My Apps

The structure for the response in these cases differs slightly, with the response for `MyApps` offering more details. The hope is that GHL will standardize the response for both these endpoints at some point.

**2) What has changed?**

* Updated code to use the full structure for the `GHLAppModel`
* Updated code to call the appropriate `detail` api for the 2 types of apps

**3) Any concern?**

None

**4) Links to design/ other documentation**

N/A

**5) Testing**

Manual testing of the UI