Code for authorization code flow for selected/ all locations
**1) What is being done?**

Once the user is on the `add-app` page, they can select 1 or more locations and click `Selected Locations` or click `All Locations` to trigger the authorization code flow for those.

The app will complete the authorization flow for the locations and also make a GET call to the redirect URLs for each of them.

**2) What has changed?**

* Updated code to allow complete authorization code flow for locations

**3) Any concern?**

None

**4) Links to design/ other documentation**

N/A

**5) Testing**

Manual testing of the UI